# flyio-admin/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "flyio-admin API",
    "description": "Application and Orchestration layer API for Flyio.ai — CMS, Blog, Trips, Knowledge Base, Web Scraper coordination, and LLM Workflow Orchestration.",
    "version": "1.0.0",
    "contact": {
      "name": "Flyio.ai Engineering"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Admin JWT token acquired via POST /api/admin/login"
      }
    },
...
  },
  "paths": {
...
    "/api/admin/blog/categories": {
      "get": {
        "summary": "List blog categories",
        "tags": ["Blog Taxonomy"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Categories list" } }
      },
      "post": {
        "summary": "Create category",
        "tags": ["Blog Taxonomy"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Category created" } }
      }
    },
    "/api/admin/blog/tags": {
      "get": {
        "summary": "List blog tags",
        "tags": ["Blog Taxonomy"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Tags list" } }
      },
      "post": {
        "summary": "Create tag",
        "tags": ["Blog Taxonomy"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Tag created" } }
      }
    },
    "/api/admin/config/ckeditor": {
      "get": {
        "summary": "Dynamic CKEditor toolbar configuration from environment",
        "tags": ["Configuration"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Toolbar items array" } }
      }
    }
  }
}